+++ db/SQL Server/2.x/2.0/src/02.functions-and-logic/sales.settle_customer_due.sql
@@ -3,7 +3,7 @@ DROP PROCEDURE sales.settle_customer_due;
 
 GO
 
-CREATE PROCEDURE sales.settle_customer_due(@customer_id bigint, @office_id integer)
+CREATE PROCEDURE sales.settle_customer_due(@customer_id integer, @office_id integer)
 AS
 BEGIN
     SET NOCOUNT ON;
